delete split modules
