--- tests/test_base.py
@@ -1,63 +1,0 @@
-"""
-Common test utilities and base classes
-"""
-
-# pylint: disable=duplicate-code
-import os
-import logging
-from unittest import TestCase
-from wsgi import app
-from service.common import status
-from service.wishlist import db, Wishlist
-from tests.factories import WishlistFactory
-
-DATABASE_URI = os.getenv(
-    "DATABASE_URI", "postgresql+psycopg://postgres:postgres@localhost:5432/testdb"
-)
-
-BASE_URL = "/wishlists"
-
-
-class BaseTestCase(TestCase):
-    """Base test case with common setup and utilities"""
-
-    @classmethod
-    def setUpClass(cls):
-        """Run once before all tests"""
-        app.config["TESTING"] = True
-        app.config["DEBUG"] = False
-        # Set up the test database
-        app.config["SQLALCHEMY_DATABASE_URI"] = DATABASE_URI
-        app.logger.setLevel(logging.CRITICAL)
-        app.app_context().push()
-
-    @classmethod
-    def tearDownClass(cls):
-        """Run once after all tests"""
-        db.session.close()
-
-    def setUp(self):
-        """Runs before each test"""
-        self.client = app.test_client()
-        db.session.query(Wishlist).delete()  # clean up the last tests
-        db.session.commit()
-
-    def tearDown(self):
-        """This runs after each test"""
-        db.session.remove()
-
-    def _create_wishlists(self, count):
-        """Factory method to create wishlists in bulk via API"""
-        wishlists = []
-        for _ in range(count):
-            wishlist = WishlistFactory()
-            resp = self.client.post(BASE_URL, json=wishlist.serialize())
-            self.assertEqual(
-                resp.status_code,
-                status.HTTP_201_CREATED,
-                "Could not create test Wishlist",
-            )
-            new_wishlist = resp.get_json()
-            wishlist.id = new_wishlist["id"]
-            wishlists.append(wishlist)
-        return wishlists
