+++ tests/test_routes.py
@@ -917,3 +917,223 @@ class TestWishlistService(TestCase):
         self.assertGreaterEqual(len(data), 1)
         for item in data:
             self.assertFalse(item["is_public"])
+
+
+######################################################################
+#  T E S T   S A D   P A T H S
+######################################################################
+class TestSadPaths(TestCase):
+    """Test REST Exception Handling"""
+
+    def setUp(self):
+        """Runs before each test"""
+        self.client = app.test_client()
+
+    def test_create_wishlist_no_data(self):
+        """It should not Create a Wishlist with missing data"""
+        response = self.client.post(BASE_URL, json={})
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_create_wishlist_no_content_type(self):
+        """It should not Create a Wishlist with no content type"""
+        response = self.client.post(BASE_URL)
+        self.assertEqual(response.status_code, status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)
+
+    def test_create_wishlist_wrong_content_type(self):
+        """It should not Create a Wishlist with the wrong content type"""
+        response = self.client.post(BASE_URL, data="hello", content_type="text/html")
+        self.assertEqual(response.status_code, status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)
+
+    def test_create_wishlist_bad_is_public(self):
+        """It should not Create a Wishlist with bad is_public data"""
+        test_wishlist = WishlistFactory()
+        # change is_public to a string
+        test_wishlist.is_public = "true"
+        response = self.client.post(BASE_URL, json=test_wishlist.serialize())
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_create_wishlist_missing_name(self):
+        """It should not Create a Wishlist without name"""
+        wishlist_data = {"customer_id": "customer123", "description": "Missing name"}
+        response = self.client.post(BASE_URL, json=wishlist_data)
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_create_wishlist_missing_customer_id(self):
+        """It should not Create a Wishlist without customer_id"""
+        wishlist_data = {"name": "Test Wishlist", "description": "Missing customer_id"}
+        response = self.client.post(BASE_URL, json=wishlist_data)
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+    def test_update_item_not_found(self):
+        """It should return 404 if the item does not exist"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        resp = self.client.put(
+            f"/wishlists/{wishlist.id}/items/99999",
+            json={"product_name": "test"},
+            content_type="application/json",
+        )
+        self.assertEqual(resp.status_code, 404)
+        self.assertIn(b"could not be found", resp.data)
+
+    def test_update_item_wrong_wishlist(self):
+        """It should return 404 if the item does not belong to the wishlist"""
+        wishlist1 = WishlistFactory()
+        wishlist1.create()
+        wishlist2 = WishlistFactory()
+        wishlist2.create()
+        item = WishlistItemFactory(wishlist=wishlist1)
+        item.create()
+        resp = self.client.put(
+            f"/wishlists/{wishlist2.id}/items/{item.id}",
+            json=item.serialize(),
+            content_type="application/json",
+        )
+        self.assertEqual(resp.status_code, 404)
+        self.assertIn(b"could not be found in Wishlist", resp.data)
+
+    def test_update_item_content_type(self):
+        """It should return 415 if Content-Type is not application/json"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        item = WishlistItemFactory(wishlist=wishlist)
+        item.create()
+        resp = self.client.put(
+            f"/wishlists/{wishlist.id}/items/{item.id}",
+            data="not json",
+            content_type="text/plain",
+        )
+        self.assertEqual(resp.status_code, 415)
+        self.assertIn(b"Content-Type must be application/json", resp.data)
+
+    def test_update_wishlist_no_content_type(self):
+        """It should not update a wishlist without content type"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        resp = self.client.put(f"{BASE_URL}/{wishlist.id}")
+        self.assertEqual(resp.status_code, status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)
+
+    def test_update_nonexistent_wishlist(self):
+        """It should return 404 when updating a non-existent wishlist"""
+        response = self.client.put(
+            f"{BASE_URL}/99999",
+            json={"name": "Test", "customer_id": "test"},
+            content_type="application/json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+    def test_add_item_no_content_type(self):
+        """It should return 415 when adding item without content type"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        response = self.client.post(f"{BASE_URL}/{wishlist.id}/items")
+        self.assertEqual(response.status_code, status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)
+
+    def test_get_wishlist_item_wrong_wishlist(self):
+        """It should return 404 when getting item from wrong wishlist"""
+        # check wishlist_id match
+        response = self.client.get(f"{BASE_URL}/99999/items/99999")
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+    def test_publish_wishlist(self):
+        """It should publish a wishlist (set is_public to True)"""
+        wishlist = WishlistFactory(is_public=False)
+        wishlist.create()
+        resp = self.client.post(f"{BASE_URL}/{wishlist.id}/publish")
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+        data = resp.get_json()
+        self.assertTrue(data["is_public"])
+        self.assertIn("published", data["message"])
+
+    def test_unpublish_wishlist(self):
+        """It should unpublish a wishlist (set is_public to False)"""
+        wishlist = WishlistFactory(is_public=True)
+        wishlist.create()
+        resp = self.client.post(f"{BASE_URL}/{wishlist.id}/unpublish")
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+        data = resp.get_json()
+        self.assertFalse(data["is_public"])
+        self.assertIn("unpublished", data["message"])
+
+    def test_publish_wishlist_not_found(self):
+        """It should return 404 when publishing a non-existent wishlist"""
+        resp = self.client.post(f"{BASE_URL}/99999/publish")
+        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertIn("could not be found", resp.get_data(as_text=True))
+
+    def test_unpublish_wishlist_not_found(self):
+        """It should return 404 when unpublishing a non-existent wishlist"""
+        resp = self.client.post(f"{BASE_URL}/99999/unpublish")
+        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertIn("could not be found", resp.get_data(as_text=True))
+
+    def test_like_wishlist_item(self):
+        """It should like a wishlist item (increment likes)"""
+        # Create a wishlist and add an item
+        wishlist = WishlistFactory()
+        wishlist.create()
+        item = WishlistItemFactory(wishlist=wishlist, likes=0)
+        item.create()
+
+        # Like the item
+        resp = self.client.post(f"{BASE_URL}/{wishlist.id}/items/{item.id}/like")
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+        data = resp.get_json()
+        self.assertEqual(data["item_id"], item.id)
+        self.assertEqual(data["wishlist_id"], wishlist.id)
+        self.assertEqual(data["likes"], 1)
+        self.assertIn("liked", data["message"])
+
+        # Like again to check increment
+        resp = self.client.post(f"{BASE_URL}/{wishlist.id}/items/{item.id}/like")
+        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+        data = resp.get_json()
+        self.assertEqual(data["likes"], 2)
+
+    def test_copy_wishlist(self):
+        """It should copy a wishlist and all its items"""
+        # Create a wishlist with items
+        wishlist = WishlistFactory()
+        wishlist.create()
+        item1 = WishlistItemFactory(wishlist=wishlist)
+        item1.create()
+        item2 = WishlistItemFactory(wishlist=wishlist)
+        item2.create()
+
+        # Copy the wishlist
+        resp = self.client.post(f"{BASE_URL}/{wishlist.id}/copy")
+        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)
+        data = resp.get_json()
+        self.assertIn("copied", data["message"])
+        self.assertNotEqual(data["original_wishlist_id"], data["new_wishlist_id"])
+        self.assertEqual(len(data["wishlist"]["wishlist_items"]), 2)
+        # Ensure copied items are not the same id as original
+        original_item_ids = {item1.id, item2.id}
+        copied_item_ids = {item["id"] for item in data["wishlist"]["wishlist_items"]}
+        self.assertTrue(copied_item_ids.isdisjoint(original_item_ids))
+
+    def test_copy_wishlist_not_found(self):
+        """It should return 404 when copying a non-existent wishlist"""
+        resp = self.client.post(f"{BASE_URL}/99999/copy")
+        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertIn("could not be found", resp.get_data(as_text=True))
+
+    def test_like_wishlist_item_not_found(self):
+        """It should return 404 when liking a non-existent item"""
+        wishlist = WishlistFactory()
+        wishlist.create()
+        resp = self.client.post(f"{BASE_URL}/{wishlist.id}/items/99999/like")
+        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertIn("could not be found", resp.get_data(as_text=True))
+
+    def test_like_wishlist_item_wrong_wishlist(self):
+        """It should return 404 when liking an item not in the wishlist"""
+        wishlist1 = WishlistFactory()
+        wishlist1.create()
+        wishlist2 = WishlistFactory()
+        wishlist2.create()
+        item = WishlistItemFactory(wishlist=wishlist1)
+        item.create()
+        resp = self.client.post(f"{BASE_URL}/{wishlist2.id}/items/{item.id}/like")
+        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertIn("could not be found in Wishlist", resp.get_data(as_text=True))
